+++ CHANGELOG.md
@@ -22,6 +22,9 @@ The format will be:
 - Introduced the `language` utility sub-router namespace under the `/v1` prefix.
 - **language**: Added a `/dictionary` endpoint to query words via an external API.
 
+### Changed
+- Added new dependencies and updated existing ones.
+
 ### Fixed
 - Updated the favicon's location to an absolute file path.
 - Fixed an error where requests weren't being logged in the case of unsuccessful requests.